fix(notices): der Owner braucht eine Entscheidung, keine Statistik
"Diese Karte sollte in der Form nicht mehr im Chat sein. Zu viel info..
bzw ich weiss nicht was ich dazu machen soll." - Owner, 2026-08-30, ueber
den taeglichen "Ziel vs. Budget"-Block der PM.

Jede automatische Meldung beantwortet jetzt zuerst EINE Frage: gibt es
darin eine Entscheidung, die nur der Owner treffen kann?

  ja   -> eine kurze Zeile in seinem Chat, bei echten Wahlmoeglichkeiten
          mit echten Buttons (pm_comm._ask_owner ueber den bestehenden
          <helmdeck-ask>-Kanal).
  nein -> Dashboard-Feed und/oder Henry (pm_comm._to_henry), der Haende
          hat und die Sache selbst erledigt. Nicht in den Owner-Chat.

Aus dem Chat verschwunden: Quota-Projektion, Plan-Gate, Dreieck-Tilt,
Kontext-Watchdog (der war an HENRY adressiert - "Karte kompaktieren,
aufteilen oder abschliessen" - und landete beim Owner), "ich kuemmere
mich", "wieder frei", "N Karten eingeplant", Quota-Hold. Der komplette
"Ziel vs. Budget"-Report laeuft weiter, aber ins Dashboard.

Neu mit Buttons: Karte ueber Budget (Stoppen / Weiterlaufen / Ansehen),
Ziel-in-Gefahr vor dem Quota-Reset (Zurueckstellen / Slip akzeptieren).

Die Laengengrenze steht in spine/comms/notice.py (zwei kurze Saetze), im
SCHREIBER erzwungen statt als Stilbitte - beide Zellen, die ungefragt in
den Chat sprechen (pm + Henry), gehen da durch, deshalb spine und nicht
eine Zelle. Der Satz-Splitter musste zweimal nachgeschaerft werden: "~6.0%"
und "So 30.08. 20:00" wurden als Satzende gelesen, die Zeile also genau
hinter der Zahl abgeschnitten und die Frage weggeworfen. Beide Faelle
sind in ops/tests/test_notice_routing.py festgenagelt, zusammen mit der
Routing-Tabelle und dem Beweis, dass die Buttons wirklich antwortbar sind.

+++ cells/copilot/henry_broker.py
@@ -578,9 +578,19 @@ def _notify_owner(text, t):
     # reads while watching Henry work, the Henry chat, showed none of the
     # broker's decisions. copilot.say is the same line the lane pipeline's
     # _say_card uses; best-effort like the push.
+    #
+    # Clipped to the same two-sentence ceiling as every other automatic notice
+    # (owner decree 2026-08-30). Henry is an LLM writing free prose into the one
+    # chat the owner actually reads, so he is the likeliest future source of the
+    # wall the decree bans - the cap belongs at this edge, not in his mandate,
+    # where it would be a request rather than a rule. Only the CHAT copy is cut:
+    # the push has always had its own 230-char budget above, and the card audit
+    # below still gets the message WHOLE, so nothing is lost, it just stops
+    # being shouted (the same split _notify_owner's docstring already draws).
     try:
         from cells.copilot import copilot
-        copilot.say(text, cls="pm", card=(t or {}).get("id") or None)
+        from spine.comms.notice import short
+        copilot.say(short(text), cls="pm", card=(t or {}).get("id") or None)
     except Exception:
         pass
     if t:
